custom-elements/instantiating-custom-elements/prototype-is-interface-prototype-object.html

<!DOCTYPE html>
<html>
<head>
<title>Prototype is already an interface prototype object</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="If PROTOTYPE is already an interface prototype object for any interface object, throw a NotSupportedError and stop">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#instantiating-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    
    var thrown = false;
    
    try {
        var XFoo = doc.register('x-foo', {prototype: HTMLElement.prototype});
    } catch (e) {
        thrown = true;
        assert_equals(e.name, 'NotSupportedError', 'Wrong exception type');
    }

    assert_true(thrown, 'Exception was not thrown');
    
}, 'Prototype is already an interface prototype object');

</script>
</body>
</html>
